pyproject.toml
[project]
name = "atmob-pillow"
version = "0.1.0"
description = "MVP image batch resize tool (PySide6 + Pillow)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyside6>=6.6",
  "pillow>=10.0",
]

[project.scripts]
image-resizer = "atmob_pillow.main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/atmob_pillow"]
